CHANGED:
Slow Area button is hidden by default and only made visible once the player has placed a gunship.

structure_select_menu.gd:

Added slow_area_unlocked to track whether or not the player can place slow areas.

The slow area button is hidden in ready().

Added unlock_slow_area to make the button visible and set the bool to true. This is called by GameManager when the player places a gunship.

GameManager.gd:

Updated unhandled_input to call unlock_slow_area() in the structure_select_menu when the player places a gunship for the first time.