ADDED:
starry_background_shader.gdshader

Ripped a starry background shader from godotshaders.com and added it as a material override to the background in our test scene.

I haven't profiled this at all, it might be very expensive in which case we will substitute with a cheaper shader or with a gif recording of the shader.